Pipeline/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
elsapy==0.5.1
idna==3.10
numpy==2.2.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
python-dotenv==1.0.1
litellm==1.55.12
pydantic[email]
matplotlib
numbers-parser